requirements.txt
flask==3.0.0
google-cloud-bigquery==3.12.0
google-cloud-aiplatform>=1.40.0
vertexai>=0.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
